+++ scripts/lichessTools-init.js
@@ -127,6 +127,7 @@ lichessTools.loadTool(LiChessTools.Tools.CevalCache);
 lichessTools.loadTool(LiChessTools.Tools.ShowKmaps);
 lichessTools.loadTool(LiChessTools.Tools.ChallengeOptions);
 lichessTools.loadTool(LiChessTools.Tools.BetterBestArrow);
+lichessTools.loadTool(LiChessTools.Tools.DownloadBroadcastPgn);
 
 
 lichessTools.loadTool(LiChessTools.Tools.HideBoardCommand);